Se refactoriza elc[odigo para utilizar el patron strategy para generar los movimientos de depositos o retios
